# docs/openapi.yaml
openapi: 3.0.3
info:
  title: dns-guard API
  version: 1.3.0
  description: |
    Local control API for the dns-guard encrypted-DNS proxy. Listen on
    127.0.0.1 (port 8090 by default; DNS_GUARD_HTTP_PORT to override,
    DNS_GUARD_HTTP=0 to disable). Same capabilities as the gRPC service
    on the unix socket.

    Typical agent flow: GET /api/v1/status → check `running`; if not,
    POST /api/v1/start; PUT /api/v1/policy to block trackers; subscribe
    to /api/v1/queries/stream (SSE) to watch live resolutions; GET
    /api/v1/stats for aggregates.
servers:
  - url: http://127.0.0.1:8090
paths:
...
  /api/v1/start:
    post:
      summary: Start the proxy (requires prior password authentication via gRPC)
      operationId: startProxy
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                mode: { type: string, enum: [doh, dot] }
                provider: { type: string, enum: [cloudflare, google, quad9] }
                strategy: { type: string, enum: [single, round-robin, failover] }
      responses:
        "200":
          description: Start result
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  message: { type: string }
...
  /api/v1/config:
    post:
      summary: Hot-swap mode/provider/strategy without restarting
      operationId: setConfig
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                mode: { type: string, enum: [doh, dot] }
                provider: { type: string, enum: [cloudflare, google, quad9] }
                strategy: { type: string, enum: [single, round-robin, failover] }
      responses:
        "200":
          description: Config result
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
  /api/v1/policy:
    get:
      summary: List the block/allow policy
      operationId: getPolicy
      responses:
        "200":
          description: Policy rules
          content:
            application/json:
              schema:
                type: object
                properties:
                  rules:
                    type: array
                    items:
                      $ref: "#/components/schemas/Rule"
    put:
      summary: Replace the block/allow policy (hot-swaps live)
      operationId: setPolicy
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                rules:
                  type: array
                  items:
                    $ref: "#/components/schemas/Rule"
      responses:
        "200":
          description: Policy result
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  message: { type: string }
...
components:
  schemas:
    Rule:
      type: object
      properties:
        pattern:
          type: string
          description: >
            Domain pattern. "example.com" matches the domain and all
            subdomains; a leading dot is ignored. Allow rules override
            block rules; the default is allow.
        action:
          type: string
          enum: [allow, block]
...